fix: remove examples and ignore test, delete shim
